minor edits

+++ sensitivity_GUI.py
@@ -114,7 +114,8 @@ class sensitivityAnalysis(object):
         fileLoc = dir_path + "//" + "infrastructures_inputs.txt"
         temp_file = dir_path + "//" + "infrastructures_inputs_temp.txt"
         files = [('Text Document', '*.txt'), ('JSON File', '*.json')] 
-        filename = askopenfilename(filetypes = files, defaultextension = files)
+        #filename = askopenfilename(filetypes = files, defaultextension = files)
+        filename = fileLoc
         copyfile(filename, temp_file)
         json_data = open(filename)
         data = json.load(json_data)
@@ -174,7 +175,8 @@ class sensitivityAnalysis(object):
         p = ggplot(results, aes(x='Value', y='RT', color = 'Sector')) + xlab(self.parameter) + ylab("Recovery Time (days)") + geom_point() + geom_line() + ggtitle(ggt)
         file_name = self.parameter + "_" + self.sector
         path_name = dir_path + "\\Sensitivity\\"
-        shutil.rmtree(path_name)
+        if os.path.isdir(path_name):
+          shutil.rmtree(path_name)
         os.mkdir(path_name)
         p.save(filename=file_name, path = path_name, verbose = False, device= "jpeg")
         copyfile(temp_file, fileLoc)
@@ -184,7 +186,7 @@ class sensitivityAnalysis(object):
         pdf.add_page()
         pdf.set_font('Times', 'B', 12)
         graph_name = path_name + file_name + ".png"
-        pdf.image(graph_name, width = 120)
+        pdf.image(graph_name, w = 200)
         text = "Slope"
         if self.parameter == "Days Backup":
           text = "Reduction (days) in recovery time for an increase in 1 day backup"
